cmake: fix tests/auto/gui/kernel/qguieventloop
The include does not join the required libraries. When building the
tests in a separate build directory this works for some reason (that's
why the CI doesn't catch it). Building everything top level breaks
though.

Considering this happens in maybe two places, I'm not sure it's worth
the effort of fixing the porting scripts.

Change-Id: I104ab9717257cbe8dfd5112dffd0d0b002cdb09e
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Qt CMake Build Bot