Fix PlatformHeaders to be a HEADER_MODULE
The module does not have an associated library, but for some reason
on Windows install(EXPORT) still created Targets files which
expected a library to be present.
If then find_package(Qt6PlatformHeaders) was called, it would fail
configuration with not finding the library on disk.

Switch the module to be a HEADER_MODULE.

Also remove the extend_target call, because there apparently is no
need to specify the headers one by one (and also because adding
sources to an interface only library would break the configuration
step).
syncqt will be executed on all the headers under the module source
location, regardless of specifying them in the project.

Change-Id: Iec4cc8fab8906e6288dae423c878e260dd922ce5
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Qt CMake Build Bot