Fix minor errors in ISODate and RFC2822Date formats' documentation
Various format characters are not capitalized in the format string, so
should not be capitalized in the doc that describes them - especially,
the m for minute should not be made to coincide with the M for month.
The H for hours *is* capitalized for ISODate, correctly indicating
that it's on a 24-hour clock, so make that explicit; and correct the
RFC2822Date doc to use HH rather than hh for its hours, likewise.
Reflow to our standard 80-column limit for docs.

Pick-to: 5.15
Change-Id: I357252817d59436de5806f45d40030b873ca3926
Reviewed-by: Paul Wicking <paul.wicking@qt.io>