Support MaximizeUsingFullscreenGeometryHint in resizeMaximizedWindows()
In QPlatformScreen there is a convenience function which resizes windows
when the screen changes. This did not have support for
MaximizeUsingFullscreenGeometryHint and would mistakenly resize these windows
to the available geometry.

Since not all QPA plugins support this hint, we have to add a capability
flag to avoid changing behavior on platforms where it works as intended.

Task-number: QTBUG-74202
Change-Id: Ife88f597fbb3affa722f63ac18fb5719ffa8ed33
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>