Use default QTD font size for mono font when importing markdown
QFontDatabase::systemFont(FixedFont) determines the font for inline code and
code blocks in a markdown document.  Now we change the size of that font to the
same size as QTextDocument::defaultFont() so that the user has the ability to
customize the font size in each document instead of only system-wide.

Change-Id: Ief7367336f7613e88695dbb08bcb7e9f50db8961
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>