eglfs: kms: Fix hw cursor with multiple screens
We used to have the assumption that moving the cursor to an out of
range position is valid and will result in a hidden cursor.

This is apparently not the case. For example, on an RPi4 with Mesa
V3D we get lots of funny artifacts after doing drmModeMoveCursor()
to invalid positions.

To remedy this, start hiding the cursor correctly when the position
is clearly out of the screen's bounds.

Task-number: QTBUG-79924
Change-Id: I3ef7ad0ce928546399443f21452f0b6deadf8036
Reviewed-by: Andy Nichols <andy.nichols@qt.io>