Fix prefix determination for windeployqt'ed applications
Qt5Core.dll of windeployqt'ed applications is right next to the
executable, and the prefix is considered the directory where the
application is located. QLibraryInfo of a relocatable Qt5Core.dll would
return a wrong prefix (by default <app dir>/..), because it determines
the prefix with QT_CONFIGURE_LIBLOCATION_TO_PREFIX_PATH (by default "..").

We now detect whether the executable was windeployqt'ed by checking
whether Qt5Core.dll is next to the executable.

However, we must not do that for applications in QT_HOST_BINS, because
they are not windeployqt'ed and must still use the standard prefix.
We detect this case by checking whether for Qt5Core.dll exists a
corresponding Qt5Core.lib in the libdir below the detected prefix.

Fixes: QTBUG-79318
Change-Id: I1c9b971b282c6b9b19a93f1819ba8aee74be5be4
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>